Change project structure to CMake style
