+++ Galib/include/File/PathFormat.h
@@ -18,7 +18,7 @@
 
 #include <cstddef>
 
-#include "Galib/GalibNamespaceDef.h"
+#include "GalibNamespaceDef.h"
 
 namespace galib::file {
     bool FormatFolderPath(char *p_desc_path, GALIB_STD size_t kLength, bool kLinuxStyle = true);
