+++ Galib/include/GalibNamespaceDef.h
