+++ Galib/include/Minecraft/Anvil.h
@@ -22,9 +22,9 @@
 #include <memory>
 #include "nbt_tags.h"
 
-#include "Galib/GalibNamespaceDef.h"
-#include "Galib/Minecraft/MinecraftCoord.h"
-#include "Galib/Exception/MinecraftException.h"
+#include "GalibNamespaceDef.h"
+#include "Minecraft/MinecraftCoord.h"
+#include "Exception/MinecraftException.h"
 
 namespace galib::minecraft {
     // Used for 2D data caching
