+++ Galib/include/Minecraft/MinecraftCoord.h
@@ -18,9 +18,9 @@
 
 #include <cstdint>
 
-#include "Galib/GalibNamespaceDef.h"
-#include "Galib/Coord/Coord2D.h"
-#include "Galib/Coord/Coord3D.h"
+#include "GalibNamespaceDef.h"
+#include "Coord/Coord2D.h"
+#include "Coord/Coord3D.h"
 
 namespace galib::minecraft {
     // Entity coordinates: Used to store positions such as the player’s location.
